Start 1.0.3 development
New: Add API function (and basic infrastructure to allow API methods to be added) which will rebuild the buttons in the collection (for example if you modify the title of a column). The API method is "fnRebuild" and accepts one parameter (optinally) - the DataTables object instance. If no instance is given, all tables which are using ColVis will be rebuilt.