Complete rewrite of ColVis to use Buttons
The new Buttons extension for DataTables is designed to be a one stop
solution for all button requirements for DataTables, and harmonises the
interface for ColVis and TableTools.

The result is that ColVis is just a set of button definitions for
Buttons now.

There is a good chance that I will move ColVis into the Buttons repo in
future and end development in this repo.